fix the following (ignored) error, which was introduced in r5520: {{{ [ making util/ptool ] test -z "" || rm -f rm -f usage: rm [-f|-i] [-dPRrvW] file ... *** Error code 1 (ignored) }}}
git-svn-id: https://svn.code.sf.net/p/gfarm/code/gfarm_v2/trunk@5683 4a5e1ece-6a47-475c-b646-731d720b41d0

+++ makes/script.mk
@@ -33,7 +33,7 @@ script-install:
 
 script-clean:
 	-test -z "$(EXTRA_CLEAN_TARGETS)" || $(RM) -f $(EXTRA_CLEAN_TARGETS)
-	-$(RM) -f $(BUILT_SCRIPTS)
+	-test -z "$(BUILT_SCRIPTS)" || $(RM) -f $(BUILT_SCRIPTS)
 
 script-veryclean: clean private-finalize
 	-test -z "$(EXTRA_VERYCLEAN_TARGETS)" || $(RM) -f $(EXTRA_VERYCLEAN_TARGETS)
